update login func
