--- test.js
@@ -1,7 +1,0 @@
-const subsInfo = {
-    "기본형" : 2,
-    "프리미엄" : 3,
-    "기업형" : 4
-}
-
-console.log(subsInfo['기본형']);
