+++ views/login/login.ejs
@@ -1,36 +1,34 @@
 <link rel="stylesheet" href="/styles/login/login.css">
-<div class="d-flex min-vh-100 justify-content-center align-items-center">
-    <div>
+<div style="display: grid; place-items: center;">
+    <div style="max-width: 320px;">
         <div class="text-center mb-4">
-            <img src="images/anyone/eco-mascot.png" alt="로고" class="mb-3" style="height: 100px;">
+            <img src="images/anyone/eco-mascot.png" alt="로고" class="mb-3 mt-3" style="height: 100px;">
             <h1 class="h5 fw-bold">에코라이프에 오신 것을 환영합니다</h1>
         </div>
         <form id="loginForm">
             <div class="mb-3 position-relative">
                 <label for="email" class="form-label">이메일</label>
-                <input type="email" id="loginEmail" class="form-control py-2 ps-4" placeholder="✉ 이메일을 입력하세요" required>
-            </div>
-            <div class="mb-3 position-relative">
+                <input type="email" id="loginEmail" class="form-control py-2 ps-4 mb-3" placeholder="✉ 이메일을 입력하세요" required>
                 <label for="password" class="form-label">비밀번호</label>
-                <input type="password" id="loginPwd" class="form-control py-2 ps-4" placeholder="🔒︎ 비밀번호를 입력하세요" required>
+                <input type="password" id="loginPwd" class="form-control py-2 ps-4 mb-3" placeholder="🔒︎ 비밀번호를 입력하세요" required>
+                <button type="submit" class="btn btn-success w-100 custom-btn py-2">로그인</button>
             </div>
-            <button type="submit" class="btn btn-success w-100 custom-btn py-2">로그인</button>
             <div class="text-center mt-3 mb-3">
                 <a href="/login/find" class="text-decoration-none text-secondary">아이디/비밀번호 찾기</a>
             </div>
-            <div class="d-flex justify-content-between">
-                <a href="/api/signup/google" style="width: 100%;">
+            <div class="justify-content-between">
+                <a href="/api/signup/google">
                     <button id="google-login" type="button" class="mt-1 btn w-100" style="border: 1px solid rgb(207, 207, 207); background: none;">
                         <img src="images/login/google.webp" alt="Google" style="max-height: 25px; object-fit: cover;">
                         Sign up with Google
                     </button>
                 </a>
-            </div>
-            <div class="d-flex justify-content-between">
-                <button  onclick="handleNaverLogin(event)" type="button" class="mt-3 btn w-100" style="border: 1px solid rgb(207, 207, 207); background: none;">
-                    <img src="images/login/naverLogo.png" alt="네이버 로그인" style="max-height: 25px; object-fit: cover;">
-                    Sign up with Naver
-                </button>
+                <a href="/api/auth/naver?server=node">
+                    <button type="button" class="mt-3 btn w-100" style="border: 1px solid rgb(207, 207, 207); background: none;">
+                        <img src="images/login/naverLogo.png" alt="네이버 로그인" style="max-height: 25px; object-fit: cover;">
+                        Sign up with Naver
+                    </button>
+                </a>
             </div>
         </form>
         <div class="text-center mt-4">
@@ -46,18 +44,5 @@
         </footer>
     </div>
 </div>
+
 <%-include('./loginConnDB.ejs') %>
-<script type="module">
-const handleNaverLogin = (event) => {
-event.preventDefault()  // 이벤트 전이 막기
-try {
-    const naverAuthUrl = 'https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=qQ8VvLEMNDDPFbFMd263&redirect_uri=https://localhost:5678/api/auth/naver/callback&state=STATE_STRING';
-    window.location.href = naverAuthUrl  // 브라우저 리다이렉션 (axios, request, fetch) 불필요
-} catch (error) {
-    console.error("인증코드 받기 실패!!", error)
-}
-}
-// 두번째 주의사항 모듈 지역변수 취급되므로 호출이 안됨
-// 해결방안은 전역변수에 등록할 것
-window.handleNaverLogin = handleNaverLogin; // 전역변수화
-</script>
