+++ views/main/header.ejs
@@ -53,7 +53,7 @@
             })
             const result = await response.json()
 
-            if(result.user.isAuthenticated) {
+            if(result.auth) {
                 $('#login, #logout, #mypage, #subs').toggle()
             }
         } catch (error) {
